bump to 2.0 with a new api
This new api doesn't require to maintain a state and pass it over different process which appear to be difficult when
you're playing with metrics. Instead at the very beginning a module is built to directly send update to the backend.

Also remove any get_value call, it is handled directly with the backend most of the time especially with remote storage.
While i'm here starts to add tests.